+++ sumo-logic/release_notes.md
@@ -1,5 +1,5 @@
 ## What's Fixed
 ### Added Operations:
-- Retrieve the List of Insights by Query
-- Retrieve the List of All Insights
-- Retrieve Details by Insights ID
+- Get the List of Insights By Query
+- Get the List of All Insights
+- Get Details By Insights ID